Move contact job cleanup to scheduled task module
