+++ hrm-domain/hrm-service/tsconfig.build.json
@@ -16,6 +16,7 @@
     { "path": "resources-domain/resources-service" },
     { "path": "hrm-domain/hrm-core" },
     { "path": "hrm-domain/scheduled-tasks/hrm-data-pull" },
+    { "path": "hrm-domain/scheduled-tasks/contact-job-cleanup" },
     { "path": "hrm-domain/hrm-service" }
   ]
 }
