+++ hrm-domain/scheduled-tasks/contact-job-cleanup/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@tech-matters/hrm-data-pull",
+  "name": "@tech-matters/contact-job-cleanup",
   "version": "1.0.0",
   "description": "The easiest way to run the tests locally is to run the test DB in docker.",
   "scripts": {
@@ -8,9 +8,10 @@
     "test": "npm run test:unit"
   },
   "dependencies" : {
-    "@tech-matters/s3-client": "^1.0.0",
     "@tech-matters/hrm-core": "^1.0.0",
     "@tech-matters/ssm-cache": "^1.0.0",
+    "@tech-matters/twilio-client": "^1.0.0",
+    "@tech-matters/types": "^1.0.0",
     "date-fns": "^2.28.0"
   }
 }
