+++ tsconfig.json
@@ -21,6 +21,7 @@
     { "path": "hrm-domain/lambdas/files-urls" },
     { "path": "hrm-domain/hrm-core" },
     { "path": "hrm-domain/scheduled-tasks/hrm-data-pull" },
+    { "path": "hrm-domain/scheduled-tasks/contact-job-cleanup" },
     { "path": "resources-domain/packages/resources-search-config" },
     { "path": "resources-domain/resources-service/tsconfig.scripts.json" },
     { "path": "resources-domain/lambdas/import-consumer" },
